bacterial-classifier/requirements.txt
tensorflow>=2.10.0
scikit-learn>=1.0.0
numpy>=1.21.0
pillow>=9.0.0
scipy>=1.7.0
streamlit>=1.10.0
pandas>=1.3.0
